+++ src/store/app.module.js
@@ -1,5 +1,10 @@
 const initialState = {
   screenOrientation: screen.orientation.type,
+  toast: {
+    message: "",
+    duration: 2000,
+    isOpen: false,
+  },
 };
 
 export const app = {
@@ -8,6 +13,22 @@ export const app = {
   mutations: {
     setScreenOrinetation(state, value) {
       state.screenOrientation = value;
+    },
+    setToastMessage(state, value) {
+      state.toast.message = value;
+    },
+    setToastDuration(state, value) {
+      state.toast.duration = value;
+    },
+    setToastState(state, value) {
+      state.toast.isOpen = value;
+    },
+  },
+  actions: {
+    showToast: async ({ commit }, payload) => {
+      commit("setToastMessage", payload.message);
+      commit("setToastDuration", payload.duration || 2000);
+      commit("setToastState", true);
     },
   },
 };