+++ src/verinfo.js
@@ -1,3 +1,3 @@
 export default {
-  version: "1.0.7",
+  version: "1.0.8",
 };