+++ CHANGELOG.md
@@ -38,3 +38,11 @@ Important Android stuff:
 - Made the whole app more consistent,
 - Set Youtube quality on "Open URL" dialog.
 - Scroll to top added to File browser
+
+## 1.0.8 (Not released yet)
+
+- API connection handling works way better,
+- Added Android notification to control playback (It's an experimental feauture should be enabled manually, you can find the option on settings. Please report any issues you have)
+- Minor optimizations,
+- Created a basic clipart icon with Android studio
+- Toast rendering improved,